+++ relay-dynamic-config/src/global.rs
@@ -1,14 +1,19 @@
 use std::path::{Path, PathBuf};
 
+use relay_event_normalization::MeasurementsConfig;
 use serde::{Deserialize, Serialize};
 
 /// A dynamic configuration for all Relays passed down from Sentry.
 ///
 /// Values shared across all projects may also be included here, to keep
 /// [`ProjectConfig`](crate::ProjectConfig)s small.
-#[derive(Default, Clone, Debug, Serialize, Deserialize)]
+#[derive(Default, Clone, Debug, Serialize, Deserialize, PartialEq)]
 #[serde(default, rename_all = "camelCase")]
-pub struct GlobalConfig {}
+#[cfg_attr(feature = "jsonschema", derive(JsonSchema))]
+pub struct GlobalConfig {
+    /// Configuration for measurements normalization.
+    pub measurements: Option<MeasurementsConfig>,
+}
 
 impl GlobalConfig {
     /// The full filename of the global config file, including the file extension.
@@ -29,5 +34,36 @@ impl GlobalConfig {
             }
             false => Ok(None),
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+
+    use relay_base_schema::metrics::MetricUnit;
+    use relay_event_normalization::{BuiltinMeasurementKey, MeasurementsConfig};
+
+    use super::GlobalConfig;
+
+    #[test]
+    fn test_global_config_roundtrip() {
+        let global_config = GlobalConfig {
+            measurements: Some(MeasurementsConfig {
+                builtin_measurements: vec![
+                    BuiltinMeasurementKey::new("foo", MetricUnit::None),
+                    BuiltinMeasurementKey::new("bar", MetricUnit::None),
+                    BuiltinMeasurementKey::new("baz", MetricUnit::None),
+                ],
+                max_custom_measurements: 5,
+            }),
+        };
+
+        let serialized =
+            serde_json::to_string(&global_config).expect("failed to serialize GlobalConfig");
+
+        let deserialized = serde_json::from_str::<GlobalConfig>(serialized.as_str())
+            .expect("failed to deserialize GlobalConfig");
+
+        assert_eq!(deserialized, global_config);
     }
 }